Fixed the packageName regular expression

+++ drools-framework-rest-guvnor/src/main/java/org/chtijbug/drools/guvnor/rest/AssetManager.java
@@ -191,7 +191,7 @@ class AssetManager {
                 return Integer.valueOf(from.getTitle());
             }
         });
-        return selectMax(allVersions, on(Integer.class).intValue());
+        return selectMax(allVersions, on(Integer.class));
     }
 
     public InputStream getPojoModel(String packageName) {
